Added code to do Razor2 'cf' (confidence) range checking.  I'm not sure exactly how it'll interact with RAZOR2_CHECK since that should take into account the user's min_cf setting, but ...
Remove some useless $fulltext stuff w/ Razor1 and 2 as well (why set
the variable if you're not going to use it?)

For a random 10 spam sample that I received today:

      8	T_RAZOR2_RANGE_90_100
      1	T_RAZOR2_RANGE_80_89
      1	T_RAZOR2_RANGE_00_09

So 80% had a super confidence level, 10% was very high, and 10% was
very low.

The default for users, BTW, is a min_cf setting of 1.


git-svn-id: https://svn.apache.org/repos/asf/incubator/spamassassin/trunk/rules@3968 13f79535-47bb-0310-9956-ffa450edef68