bug 1473 made the add_header option which should be used for _all_ configuration of headers.
_TAGS_ can be used with add_header and the report templates to add
interesting info into the reports

To add a tag, change the documentation in Conf.pm, and either put the
info into  $self->{tag_data}->{TAG} or add a anonymous sub ref to the
hash %tags in PerMsgStatus.pm

There is also a bit of backwards compatibility stuff involved here, although
some of it should be permanent (like default headers, maybe).

I did profile the code, and there doesn't seem to be much of a change
at all.


git-svn-id: https://svn.apache.org/repos/asf/incubator/spamassassin/trunk@5395 13f79535-47bb-0310-9956-ffa450edef68